+++ src/fakeMotor.cpp
@@ -76,8 +76,8 @@ int FakeMotor::get_error_code() { return error_code_; }
 
 double FakeMotor::get_current_position() { return current_position_; }
 
-int FakeMotor::get_current_velocity() { return current_velocity_; }
+double FakeMotor::get_current_velocity() { return current_velocity_; }
 
-int FakeMotor::get_current_torque() { return current_torque_; }
+double FakeMotor::get_current_torque() { return current_torque_; }
 
 MotorModes FakeMotor::get_mode() { return mode_; }
